.ipynb_checkpoints/requirements-checkpoint.txt
ftfy==6.3.1
matplotlib==3.10.8
mediapipe==0.10.14
numpy==1.24.3
opencv_contrib_python==4.13.0.92
opencv_python==4.8.1.78
pandas==3.0.2
peft==0.4.0
Pillow==12.2.0
PyYAML==6.0.3
regex==2026.1.15
scikit_learn==1.2.2
timm==0.6.13
torch==2.0.1
torchvision==0.15.2
tqdm==4.67.1
yacs==0.1.8
